+++ tests/testthat/test-symmetricPolynomial.R
@@ -3,8 +3,8 @@ test_that("Symmetric polynomial", {
   e2 <- ESFpoly(4, 2)
   e3 <- ESFpoly(4, 3)
   f <- e1 + 2*e2 + 3*e3
-  expect_true(isSymmetricPolynomial(f))
-  expect_false(isSymmetricPolynomial(qlone(1) - qlone(2)))
+  expect_true(isSymmetricQspray(f))
+  expect_false(isSymmetricQspray(qlone(1) - qlone(2)))
 })
 
 test_that("MSPcombination", {